delete CommonData
